refactor : rename PostContent component to Post component
