+++ components/Post/indexStyle.tsx
